+++ src/idtrackerai/extra_tools/video_generator/video_generator_gui.py
@@ -3,8 +3,8 @@ from pathlib import Path
 
 import cv2
 import numpy as np
-from qtpy.QtCore import QRect, Qt
-from qtpy.QtGui import QIcon, QImage, QPixmap
+from qtpy.QtCore import QRect, Qt, QUrl
+from qtpy.QtGui import QDesktopServices, QIcon, QImage, QPixmap
 from qtpy.QtWidgets import (
     QApplication,
     QFileDialog,
@@ -46,7 +46,7 @@ class VideoGeneratorGUI(GUIBase):
     def __init__(self) -> None:
         super().__init__()
 
-        self.setWindowTitle("idtracker.ai | Video Generator")
+        self.setWindowTitle("Video Generator")
         self.documentation_url = (
             "https://idtracker.ai/latest/user_guide/video_generators.html"
         )
@@ -262,9 +262,28 @@ class VideoGeneratorGUI(GUIBase):
         except KeyboardInterrupt:
             pass  # User cancelled the operation
         else:
-            QMessageBox.information(
-                self, "Done", f"Video generation finished, saved in {output_path}."
+            msg_box = QMessageBox(self)
+            msg_box.setIcon(QMessageBox.Icon.Information)
+            msg_box.setWindowTitle("Done")
+            msg_box.setText(f"Video generation finished, saved in {output_path}.")
+            open_folder_btn = msg_box.addButton(
+                "Open File Location", QMessageBox.ButtonRole.ActionRole
             )
+            open_folder_btn.setIcon(QIcon.fromTheme("folder-open"))
+            msg_box.addButton(QMessageBox.StandardButton.Ok)
+            msg_box.exec()
+
+            if msg_box.clickedButton() == open_folder_btn:
+                if Path(output_path).is_file():
+                    # If output_path is a file, get its directory
+                    output_path = str(Path(output_path).parent)
+                try:
+                    QDesktopServices.openUrl(QUrl.fromLocalFile(output_path))
+                except Exception as e:
+                    QMessageBox.warning(
+                        self, "Error Opening Folder", f"Could not open the folder: {e}"
+                    )
+        progress_dialog.close()
 
     def individual_size_changed(self, value: int):
         self.positions, self.individual_output_shape = get_geometries(
@@ -316,7 +335,7 @@ class VideoGeneratorGUI(GUIBase):
 
         try:
             self.trajectories = np.nan_to_num(
-                load_trajectories(self.session.session_folder)["trajectories"], nan=-1
+                load_trajectories(self.session.session_folder)["trajectories"], nan=-100
             )
         except FileNotFoundError as exc:
             QMessageBox.warning(