+++ docs/source/user_guide/changelog.rst
@@ -10,9 +10,9 @@ Authors since :ref:`5.0.0`: Jordi Torrents (jordi.torrentsm@gmail.com)
 - Fix a bug in the :ref:`validator_reference` with ``numpy.bool`` getting into PyQt system.
 - Improve main Argument Parser descriptions.
 - Fix incompatibility with non-ASCII (chinese) characters in paths.
-- Fix frozen Validator GUI when opening and saving sessions.
+- Fix frozen Validator GUI when opening and saving sessions due to bad parallel processing management.
 - Add new parameter ``CONTRASTIVE_MIN_ACCUMULATION`` to set the minimum fraction of images that need to be accumulated for taking the contrastive step as sufficient. If the fraction of accumulated images is lower than this value, the accumulation protocol will be run. This parameter is set to 0.5 by default.
-- Optimize rendering in the :ref:`validator_reference` preview.
+- Optimized object rendering in the :ref:`validator_reference` preview by checking if the object is visible in the current frame before rendering it.
 - Fix video player not respecting original video framerate.
 
 6.0.5